adding file structure import
