+++ app/todo.py
@@ -1,34 +1,8 @@
 from datetime import datetime, timedelta
 import json
-import getpass
-import hashlib
 import os
 
-#### User Registration Start Here ####
-def register():
-    username = input("Enter Your Username: ")
-    password = getpass.getpass("Enter Your Password: ")
-    hashed_password = hashlib.sha256(password.encode()).hexdigest()
 
-    users = load_users()
-    if username in users:
-        print("User already exists. Please Login.")
-        return
-    users[username] = hashed_password
-    save_users(users)
-    print("Registration successful.")
-
-def login():
-    username = input("Enter Your Username: ")
-    password = getpass.getpass("Enter Your Password: ")
-    hashed_password = hashlib.sha256(password.encode()).hexdigest()
-
-    users = load_users()
-    if username not in users or users[username] != hashed_password:
-        print("Invalid username or password.")
-        return None
-    print("Login successful.")
-    return username
 
 def load_users():
     try:
@@ -43,26 +17,6 @@ def save_users(users):
 
 def get_todo_file(username):
     return f"{username}_todo_list.json"
-
-#### User Registration End Here ####
-
-def display_menu():
-    print("======================")
-    print("To-Do List Application")
-    print("======================")
-    print("1. View To-Do List")
-    print("2. Add To-Do Item")
-    print("3. Edit To-Do Item")
-    print("4. Mark To-Do Item as Complete")
-    print("5. View Completed Item")
-    print("6. Remove To-Do Item")
-    print("7. View To-Do List by Category")
-    print("8. Search To-Do Items by keyword")
-    print("9. Sort To-Do List")
-    print("10. Set Reminders for Tasks")
-    print("11. Save To-Do List")
-    print("12. Load To-Do List")
-    print("13. Exit")
 
 def view_todo_list(username):
     print("\nTo-Do List:")
@@ -83,7 +37,6 @@ def view_todo_list(username):
                 except ValueError:
                     print(f"Error in item: {item}")
     print()
-    
 
 def add_todo_item(username):
     item = input("Enter a new to-do item: ").strip().lower()
@@ -282,57 +235,6 @@ def load_json(username):
         print("To-Do List loaded from todo_list.json")
     except FileNotFoundError:
         print("No To-Do List found. Please add items to the list first.")
-
-def main():
-    while True:
-        print("*************")
-        print("1. Register")
-        print("2. Login")
-        print("3. Exit")
-        print("*************")
-        choice = input("Enter you choice: ")
-        if choice == '1':
-            register()
-        elif choice == '2':
-            username = login()
-            if username:
-                while True:
-                    display_menu()
-                    choice = input("Enter your choice: ")
-                    if choice == '1':
-                        view_todo_list(username)
-                    elif choice == '2':
-                        add_todo_item(username)
-                    elif choice == '3':
-                        edit_todo_item(username)
-                    elif choice == '4':
-                        mark_complete(username)
-                    elif choice == '5':
-                        view_completed_items(username)
-                    elif choice == '6':
-                        remove_todo_item(username)
-                    elif choice == '7':
-                        view_by_category(username)
-                    elif choice == '8':
-                        search_todo_items(username)
-                    elif choice == '9':
-                        sort_todo_list(username)
-                    elif choice == '10':
-                        set_reminder(username)
-                    elif choice == '11':
-                        save_to_json(username)
-                    elif choice == '12':
-                        load_json(username)
-                    elif choice == '13':
-                        print('GoodBye.')
-                        break
-                    else:
-                        print("Invalid choice. Please try again.\n")
-        elif choice == '3':
-            print("GoodBye.")
-            break
-        else:
-            print("Invalid choice. Please try again.\n")
 
 if __name__ == "__main__":
     main()
