+++ app/utils.py
@@ -1,0 +1,21 @@
+
+#### To-Do List Application Start Here ####
+def display_menu():
+    print("======================")
+    print("To-Do List Application")
+    print("======================")
+    print("1. View To-Do List")
+    print("2. Add To-Do Item")
+    print("3. Edit To-Do Item")
+    print("4. Mark To-Do Item as Complete")
+    print("5. View Completed Item")
+    print("6. Remove To-Do Item")
+    print("7. View To-Do List by Category")
+    print("8. Search To-Do Items by keyword")
+    print("9. Sort To-Do List")
+    print("10. Set Reminders for Tasks")
+    print("11. Save To-Do List")
+    print("12. Load To-Do List")
+    print("13. Exit")
+
+print("======================")
